+++ packages/frontend/src/components/CustomVisualization/CustomVisualizationProvider/useCustomVisualizationContext.tsx
@@ -1,0 +1,5 @@
+import { useContext } from 'react';
+import { CustomVisualizationContext } from '.';
+
+export const useCustomVisualizationContext = () =>
+    useContext(CustomVisualizationContext);
