+++ packages/frontend/src/components/CustomVisualization/index.tsx
@@ -1,83 +1,15 @@
-import { ResultRow } from '@lightdash/common';
 import { Center, Code, Loader } from '@mantine/core';
-import {
-    createContext,
-    FC,
-    lazy,
-    Suspense,
-    useContext,
-    useMemo,
-    useState,
-} from 'react';
-import { useExplorerContext } from '../../providers/ExplorerProvider/useExplorerContext';
-
-const VegaLite = lazy(() =>
-    import('react-vega').then((module) => ({ default: module.VegaLite })),
-);
-
-const defaultValue = JSON.stringify({}, null, 2);
-
-const CustomVisualizationContext = createContext<{
-    chartConfig: string;
-    setChartConfig: (newConfig: string) => void;
-    rows: {
-        [k: string]: unknown;
-    }[];
-    fields: string[];
-}>({
-    chartConfig: defaultValue,
-    setChartConfig: () => {},
-    rows: [],
-    fields: [],
-});
-
-export const useCustomVisualizationContext = () =>
-    useContext(CustomVisualizationContext);
-
-const convertRowsToSeries = (rows: ResultRow[]) => {
-    return rows.map((row) => {
-        return Object.fromEntries(
-            Object.entries(row).map(([key, rowValue]) => [
-                key,
-                rowValue.value.raw,
-            ]),
-        );
-    });
-};
-
-export const CustomVisualizationProvider: FC = ({ children }) => {
-    const rows = useExplorerContext(
-        (context) => context.queryResults.data?.rows,
-    );
-
-    const [chartConfig, setChartConfig] = useState<string>(defaultValue);
-
-    const convertedRows = useMemo(() => {
-        return rows ? convertRowsToSeries(rows) : [];
-    }, [rows]);
-
-    const fields = useMemo(() => {
-        return rows && rows.length > 0 ? Object.keys(rows[0]) : [];
-    }, [rows]);
-
-    return (
-        <CustomVisualizationContext.Provider
-            value={{
-                chartConfig,
-                setChartConfig,
-                rows: convertedRows,
-                fields,
-            }}
-        >
-            {children}
-        </CustomVisualizationContext.Provider>
-    );
-};
+import { FC, lazy, Suspense, useMemo } from 'react';
+import { useCustomVisualizationContext } from './CustomVisualizationProvider/useCustomVisualizationContext';
 
 type CustomVisualizationProps = {
     className?: string;
     'data-testid'?: string;
 };
+
+const VegaLite = lazy(() =>
+    import('react-vega').then((module) => ({ default: module.VegaLite })),
+);
 
 const CustomVisualization: FC<CustomVisualizationProps> = (props) => {
     const { chartConfig, rows } = useCustomVisualizationContext();
