+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/CustomVisConfigTabs.tsx
@@ -2,7 +2,7 @@ import { Loader, Tabs } from '@mantine/core';
 import Editor, { EditorProps, Monaco } from '@monaco-editor/react';
 import merge from 'lodash/merge';
 import React, { memo, useEffect, useRef, useState } from 'react';
-import { useCustomVisualizationContext } from '../../CustomVisualization';
+import { useCustomVisualizationContext } from '../../CustomVisualization/CustomVisualizationProvider/useCustomVisualizationContext';
 
 type Schema = {
     readonly uri: string;
